Set all hidden area mesh types
The "LineLoop" mesh added in a previous commit appears to be used by the
compositor just before displaying the image to the display, but the
games I tested did not seem to use it to cull pixels before then.

Games can query different styles of meshes, and it seems like most of
them only check for the "Standard" mesh.