Only set Standard Hidden Area Mesh
The Line Loop hidden area mesh is used by the compositor to modify the
distortion mesh, but I've not been able to figure out how to set it
without the compositor running into issues on first boot.
So for now we just set the Standard mesh, which is the one that actually
matters for performance.

Other changes:
- Moved HAM updates to `UpdateSettings`. Restarting your game should be
  enough for you to see your updated mesh.
- Removed the black level increase from the HAM test mode (since that
  only works properly with the Line Loop mesh).